Merge some common record and playback logic, and match ISRs to priorities:
Combine buffer logic for record and playback mode since it's now identical (if waiting, schedule work).
Change the ISRs and enable flags so that the SD ISR is always the lower priority one, and make it so that none of the ISRs are triggered by the CompareA/CompareB registers (which would be affected by PWM and lead to variable timing).
Rename togByte to be more descriptive (it's the lower-priority SD ISR enable), and added a similarly-named property for the higher-priority Sample ISR)

Also noted a few potential bugs, in particular around looping/stopping when there is data still in the buffer.